Move browserHistory to a distinct module
